+++ liteiclink/serdes/gtx_7series_init.py
@@ -92,7 +92,7 @@ class GTXInit(Module):
             )
         )
         # Wait 500ns after configuration before releasing GTX reset (to follow AR43482)
-        init_delay = WaitTimer(int(500e-9*sys_clk_freq))
+        init_delay = WaitTimer(500e-9*sys_clk_freq)
         self.submodules += init_delay
         self.comb += init_delay.wait.eq(1)
         fsm.act("WAIT-INIT-DELAY",
@@ -154,7 +154,7 @@ class GTXInit(Module):
         )
 
         # FSM watchdog / restart
-        watchdog = WaitTimer(int(1e-3*sys_clk_freq))
+        watchdog = WaitTimer(1e-3*sys_clk_freq)
         self.submodules += watchdog
         self.comb += [
             watchdog.wait.eq(~fsm.reset & ~self.done),
